// src/app/module/ng-mat-list-swipe/enums/warnings.ts
export enum Warnings {
    CONFIG_NOT_LOADED = 'CONFIG_NOT_LOADED',
    ADDING_DEFAULT_SLIDE_THRESHOLD = 'ADDING_DEFAULT_SLIDE_THRESHOLD',
    ZERO_SLIDE_THRESHOLD_NOT_ALLOWED = 'ZERO_SLIDE_THRESHOLD_NOT_ALLOWED',
    SLIDE_THRESHOLD_NOT_FOUND = 'SLIDE_THRESHOLD_NOT_FOUND',
    MAX_OFFSET_EXCEEDED = 'MAX_SLIDE_THRESHOLD_NOT_ALLOWED',
    INVALID_SLIDE_THRESHOLD_NOT_ALLOWED = 'INVALID_SLIDE_THRESHOLD_NOT_ALLOWED'
}